requirements.txt
flask
flask_wtf
flask_login
werkzeug
pymysql
python-dotenv